+++ cyclops-javaslang/src/test/java/com/aol/cyclops/javaslang/streams/StreamUtilsStreamTest.java
@@ -5,6 +5,9 @@ import static java.util.Arrays.asList;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasItem;
 import static org.junit.Assert.assertThat;
+
+import java.util.List;
+
 import javaslang.collection.LazyStream;
 
 import org.junit.Test;
@@ -27,7 +30,7 @@ public class StreamUtilsStreamTest {
 	@Test
 	public void testFlatMap(){
 
-		assertThat(StreamUtils.flatMapStream(LazyStream.of( asList("1","10"), asList("2"),asList("3"),asList("4")), list -> list.stream() ).toJavaList(),hasItem("10"));
+		assertThat(StreamUtils.flatMapStream(LazyStream.<List<String>>of( asList("1","10"), asList("2"),asList("3"),asList("4")), list -> list.stream() ).toJavaList(),hasItem("10"));
 
 	}
